Refactor source files location for easier navigation
