+++ src/stalker/AddonSettings.cpp
