+++ src/stalker/AddonSettings.h
