+++ src/stalker/CWatchdog.cpp
