+++ src/stalker/CWatchdog.h
