+++ src/stalker/ChannelManager.cpp
