+++ src/stalker/ChannelManager.h
