+++ src/stalker/Error.h
