+++ src/stalker/GuideManager.cpp
