+++ src/stalker/GuideManager.h
