+++ src/stalker/HTTPSocket.cpp
