+++ src/stalker/InstanceSettings.cpp
