+++ src/stalker/InstanceSettings.h
