+++ src/stalker/SessionManager.cpp
