+++ src/stalker/SessionManager.h
