+++ src/stalker/SettingsMigration.cpp
