+++ src/stalker/SettingsMigration.h
