+++ src/stalker/Utils.cpp
