+++ src/stalker/Utils.h
