+++ src/stalker/base/Cache.cpp
