+++ src/stalker/base/Cache.h
