+++ src/stalker/base/ChannelManager.h
