+++ src/stalker/base/GuideManager.h
